Move WalnutAmmo class
